### 2023-06-11 #5537 @arfon
@PieterjanRobbe – due to the relatively small size of this code, the editors will now discuss if it meets the [substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) criterion for review by JOSS. We should get back to you sometime next week. If you want to fix the DOIs (noting that @editorialbot's suggestions are not always right), you can, then use the following commands (one at a time, as the first line of a new comment) to regenerate the PDF and check the references.

Note to editors, the line count for this submission is a little misleading as there are a couple of large data files. The Julia in `src` is:

```
Language                     files          blank        comment           code
Julia                           19            506            766            846
SUM:                            19            506            766            846
```
