+++ paper/paper.bib
@@ -19,6 +19,7 @@
 
 @article{blondeel2020,
   title   = {p-refined multilevel quasi-Monte Carlo for Galerkin finite element methods with applications in civil engineering},
+  doi     = {10.3390/a13050110},
   author  = {Blondeel, Philippe and Robbe, Pieterjan and Van hoorickx, C{\'e}dric and Fran{\c{c}}ois, Stijn and Lombaert, Geert and Vandewalle, Stefan},
   journal = {Algorithms},
   volume  = {13},
@@ -29,6 +30,7 @@
 
 @article{chiang2000,
   title   = {Phase information and the evolution of cosmological density perturbations},
+  doi     = {10.1046/j.1365-8711.2000.03086.x},
   author  = {Chiang, Lung-Yih and Coles, Peter},
   journal = {Monthly Notices of the Royal Astronomical Society},
   volume  = {311},
@@ -47,6 +49,7 @@
 
 @inproceedings{cui2018,
   title     = {Uncertainty quantification of electronic and photonic ICs with non-Gaussian correlated process variations},
+  doi       = {10.1145/3240765.3240860},
   author    = {Cui, Chunfeng and Zhang, Zheng},
   booktitle = {Proceedings of the International Conference on Computer-Aided Design},
   pages     = {1--8},
@@ -82,6 +85,7 @@
 
 @article{pirot2015,
   title   = {Influence of conceptual model uncertainty on contaminant transport forecasting in braided river aquifers},
+  doi     = {10.1016/j.jhydrol.2015.07.036},
   author  = {Pirot, Guillaume and Renard, Philippe and Huber, Emanuel and Straubhaar, Julien and Huggenberger, Peter},
   journal = {Journal of Hydrology},
   volume  = {531},
@@ -90,8 +94,8 @@
 }
 
 @article{PlotsJL,
-  doi       = {https://doi.org/10.5334/jors.431},
   url       = {https://openresearchsoftware.metajnl.com/articles/10.5334/jors.431/},
+  doi       = {10.5334/jors.431},
   author    = {Christ, Simon and Schwabeneder, Daniel and Rackauckas, Christopher and Borregaard, Michael Krabbe and Breloff, Thomas},
   keywords  = {Graphics (cs.GR), FOS: Computer and information sciences, FOS: Computer and information sciences, I.3.3},
   title     = {Plots.jl -- a user extendable plotting API for the julia programming language},
@@ -111,6 +115,7 @@
 
 @article{robbe2021,
   title   = {Enhanced multi-index Monte Carlo by means of multiple semicoarsened multigrid for anisotropic diffusion problems},
+  doi     = {10.1002/nla.2281},
   author  = {Robbe, Pieterjan and Nuyens, Dirk and Vandewalle, Stefan},
   journal = {Numerical Linear Algebra with Applications},
   volume  = {28},
@@ -121,6 +126,7 @@
 
 @article{stephenson2006,
   title   = {Adaptive Markov random fields for example-based super-resolution of faces},
+  doi     = {10.1155/asp/2006/31062},
   author  = {Stephenson, Todd A and Chen, Tsuhan},
   journal = {EURASIP Journal on Advances in Signal Processing},
   volume  = {2006},
@@ -129,11 +135,12 @@
 }
 
 @article{wu2023,
-  title={Physics-embedded inverse analysis with algorithmic differentiation for the earth’s subsurface},
-  author={Wu, Hao and Greer, Sarah Y and O’Malley, Daniel},
-  journal={Scientific Reports},
-  volume={13},
-  number={1},
-  pages={718},
-  year={2023}
[…]
