### 2023-06-12 #5537 @PieterjanRobbe
Hi @arfon, thanks for considering this submission! I've added the DOI's, and reference checking seems fine now.
