### 2023-06-13 #5537 @PieterjanRobbe
Just pointing out that this package has been used in several independent research works, including:
- Pulsipher, J. L., Benjamin R. D., and Victor M. Z. "Random field optimization." Computers & Chemical Engineering 165 (2022): 107854. ([link](https://www.sciencedirect.com/science/article/pii/S0098135422001922?casa_token=CBInJp7n91AAAAAA:5Cor6NDla79FpTi2W3KC7EEL1Nsbd-rO554IvGppQ2r-57DNBlsTD_lvlSSpCBy4COKucHwhIk4))
- Wu, H., Greer, S. Y., and 'Malley, D. "Physics-embedded inverse analysis with algorithmic differentiation for the earth’s subsurface." Scientific Reports 13.1 (2023): 718. ([link](https://www.nature.com/articles/s41598-022-26898-1))
- Wu, H., O'Malley, D., Golden, J. K., and Vesselinov, V. V. "Inverse analysis with variational autoencoders: A comparison of shallow and deep networks." Journal of Machine Learning for Modeling and Computing 3.2 (2022). ([link](https://www.dl.begellhouse.com/journals/558048804a15188a,4c1b7528188eccff,0ff0a9921c9bc98e.html))
- Schäfer, F., Katzfuss, M., and Owhadi, H. "Sparse Cholesky Factorization by Kullback-Leibler Minimization." SIAM Journal on scientific computing 43.3 (2021): A2019-A2046. ([link](https://epubs.siam.org/doi/abs/10.1137/20M1336254))
- Belliveau, P., and Haber, E. "Parametric level-set inverse problems with stochastic background estimation." Inverse Problems 39.7 (2023): 075003. ([link](https://iopscience.iop.org/article/10.1088/1361-6420/acd413/pdf))
- Revell, C. K., Herrera, J. A., Lawless, C., Lu, Y., Kadler, K. E., Chang, J., and Jensen, O. E. "Modelling collagen fibril self-assembly from extracellular medium in embryonic tendon." (preprint) ([link](https://www.biorxiv.org/content/biorxiv/early/2023/03/15/2023.03.13.532430.full.pdf))
- Hang, H., Giusti, C., Ziegelmeier, L., & Henselman-Petrusek, G. "U-match factorization: sparse homological algebra, lazy cycle representatives, and dualities in persistent (co) homology." (preprint) ([link](https://arxiv.org/pdf/2108.08831))

as well as in our own work:
- Blondeel, P., Robbe, P., Van hoorickx, C., François, S., Lombaert, G., & Vandewalle, S. "p-refined multilevel quasi-Monte Carlo for Galerkin finite element methods with applications in civil engineering." Algorithms, 13.5 (2022): 110. ([link](https://www.mdpi.com/1999-4893/13/5/110))
- Robbe, P., Nuyens, D., and Vandewalle, S. "Enhanced multi‐index Monte Carlo by means of multiple semicoarsened multigrid for anisotropic diffusion problems." Numerical Linear Algebra with Applications 28.3 (2021): e2281. ([link](https://onlinelibrary.wiley.com/doi/pdf/10.1002/nla.2281))
- Robbe, P., Nuyens, D., and Vandewalle, S. "A multi-index quasi-Monte Carlo algorithm for lognormal diffusion problems." SIAM Journal on Scientific Computing 39.5 (2017): S851-S872. ([link](https://epubs.siam.org/doi/pdf/10.1137/16M1082561))
- Robbe, P., Nuyens, D., and Vandewalle, S. "Recycling samples in the multigrid multilevel (quasi-) Monte Carlo method." SIAM Journal on Scientific Computing 41.5 (2019): S37-S60. ([link](https://epubs.siam.org/doi/pdf/10.1137/18M1194031))
