### 2023-06-26 #5537 @jbytecode
👋👋👋 Dear @shahmoradi and Dear @ziyiyin97 👋👋👋

JOSS publishes articles about open source research software. The submission I'd like you to review is titled: "**[GaussianRandomFields.jl: A Julia package to generate and sample from Gaussian random fields](https://github.com/openjournals/joss-reviews/issues/5537#top)**". You can find more information at the top of this Github issue (https://github.com/openjournals/joss-reviews/issues/5537).

Thank you in advance!
