### 2023-06-27 #5595 @editorialbot
[REVIEW]: GaussianRandomFields.jl: A Julia package to generate and sample from Gaussian random fields
**Submitting author:** @PieterjanRobbe (<a href="http://orcid.org/0000-0002-6254-8245">Pieterjan Robbe</a>)
**Repository:** https://github.com/PieterjanRobbe/GaussianRandomFields.jl
**Branch with paper.md** (empty if default branch): joss
**Version:** v2.2.4
