### 2023-07-28 PieterjanRobbe/GaussianRandomFields.jl#50 @ziyiyin97
JOSS manuscript review
This issue entails my review comments for the JOSS manuscript submission at https://github.com/openjournals/joss-reviews/issues/5595

The software is well-scoped, aiming at solving at a very concrete problem -- generating samples from Gaussian random field. It is quite adaptable and generalizable to different setups of mean and covariance functions. It plays well with other packages -- e.g. with KernelFunctions.jl and customized meshes. The plotting utilities also look great. Well done!!

The documentation is also very well-written. The tutorials are very straightforward and easy to follow, precisely showing the capabilities of the package. One minor suggestion (following the JOSS guideline) is that author(s) should put a contribution guideline to either the README or the docs (or both), indicating how to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

The paper is also succinct and straightforward. It mentioned related packages and clearly pointed out the advantage of this GaussianRandomFields.jl package. Also, this package has already been used in a few papers listed in the manuscript.

I recommend acceptance with only one suggestion about adding the contribution guidelines.
