### 2023-07-28 PieterjanRobbe/GaussianRandomFields.jl#50 @PieterjanRobbe
Thank you so much @ziyiyin97! I've added the contribution guidelines to the README, see 5bab000.
