### 2023-07-28 #5595 @PieterjanRobbe
Hi @ziyiyin97, thank you so much for your time reviewing the software and paper! I've addressed your comments with [this commit](https://github.com/PieterjanRobbe/GaussianRandomFields.jl/commit/5bab000fe1d1a4d6c8eac2c62a1096c8774adde9).
