### 2023-07-31 PieterjanRobbe/GaussianRandomFields.jl#50 @ziyiyin97
LGTM. Closing
