### 2023-08-08 086140c
Added instructions for testing to README
