### 2023-08-08 8bd9cec
Added reference for performance claim
