### 2023-08-08 #5595 @PieterjanRobbe
Hi @shahmoradi, thank you so much for your time and effort! Here's an answer to your questions:
1. If you're referring to the code snippet under [Package Overview](https://pieterjanrobbe.github.io/GaussianRandomFields.jl/dev/tutorial/#Package-Overview), then yes, this is intentional. The goal of these lines is to illustrate the general workflow of the package, where `Kernel()` and `Generator()` are supposed to be replaced by a call to the appropriate methods, as mentioned below the code block. I've tested all other code blocks in the tutorial, and they seem to work fine for me.
2. Perhaps this was not clearly stated in the manuscript. The performance benefits mentioned in the paper refer to the performance increase of Julia over R/Python (see [1]): "As such, [the package] benefits from the performance advantage of Julia, [and ...]". I've added a reference for this performance claim, but I could also rephrase/remove that sentence if needed.
3. I've added instructions for testing to the README.

Thanks again!
