### 2023-08-31 cfc21ae
update bibliography
