+++ paper/paper.bib
@@ -18,7 +18,7 @@
 }
 
 @article{blondeel2020,
-  title   = {p-refined multilevel quasi-Monte Carlo for Galerkin finite element methods with applications in civil engineering},
+  title   = {p-refined multilevel quasi-{M}onte {C}arlo for {G}alerkin finite element methods with applications in civil engineering},
   doi     = {10.3390/a13050110},
   author  = {Blondeel, Philippe and Robbe, Pieterjan and Van hoorickx, C{\'e}dric and Fran{\c{c}}ois, Stijn and Lombaert, Geert and Vandewalle, Stefan},
   journal = {Algorithms},
@@ -40,15 +40,16 @@
 }
 
 @book{chiles2012,
-  title     = {Geostatistics: modeling spatial uncertainty},
+  title     = {Geostatistics: {M}odeling spatial uncertainty},
   author    = {Chiles, Jean-Paul and Delfiner, Pierre},
   volume    = {713},
   year      = {2012},
+  doi       = {10.1016/s0098-3004(00)00063-7},
   publisher = {John Wiley \& Sons}
 }
 
 @inproceedings{cui2018,
-  title     = {Uncertainty quantification of electronic and photonic ICs with non-Gaussian correlated process variations},
+  title     = {Uncertainty quantification of electronic and photonic ICs with non-{G}aussian correlated process variations},
   doi       = {10.1145/3240765.3240860},
   author    = {Cui, Chunfeng and Zhang, Zheng},
   booktitle = {Proceedings of the International Conference on Computer-Aided Design},
@@ -58,7 +59,7 @@
 
 @article{GSTools,
   author  = {M\"uller, S. and Sch\"uler, L. and Zech, A. and He{\ss}e, F.},
-  title   = {\texttt{GSTools} v1.3: a toolbox for geostatistical modelling in Python},
+  title   = {\texttt{GSTools} v1.3: a toolbox for geostatistical modelling in {P}ython},
   journal = {Geoscientific Model Development},
   volume  = {15},
   year    = {2022},
@@ -98,7 +99,7 @@
   doi       = {10.5334/jors.431},
   author    = {Christ, Simon and Schwabeneder, Daniel and Rackauckas, Christopher and Borregaard, Michael Krabbe and Breloff, Thomas},
   keywords  = {Graphics (cs.GR), FOS: Computer and information sciences, FOS: Computer and information sciences, I.3.3},
-  title     = {Plots.jl -- a user extendable plotting API for the julia programming language},
+  title     = {Plots.jl -- {A} user extendable plotting API for the {J}ulia programming language},
   publisher = {Journal of Open Research Software},
   year      = {2023},
   copyright = {Creative Commons Attribution 4.0 International}
@@ -114,7 +115,7 @@
 }
 
 @article{robbe2021,
-  title   = {Enhanced multi-index Monte Carlo by means of multiple semicoarsened multigrid for anisotropic diffusion problems},
+  title   = {Enhanced multi-index {M}onte {C}arlo by means of multiple semicoarsened multigrid for anisotropic diffusion problems},
   doi     = {10.1002/nla.2281},
   author  = {Robbe, Pieterjan and Nuyens, Dirk and Vandewalle, Stefan},
   journal = {Numerical Linear Algebra with Applications},
