### 2023-08-31 #5595 @jbytecode
@PieterjanRobbe - I've just sent a pull request that contains minor changes in your bibtex. Could you please review the changes and apply them if they are okay?

PR: https://github.com/PieterjanRobbe/GaussianRandomFields.jl/pull/51

Thank you in advance.
