### 2023-08-31 #5595 @jbytecode
@PieterjanRobbe  - One more PR about bibtex

PR: https://github.com/PieterjanRobbe/GaussianRandomFields.jl/pull/52
