### 2023-08-31 #5595 @jbytecode
@PieterjanRobbe - Meanwhile checking out the references and the manuscript, you can follow the instructions given in the https://github.com/openjournals/joss-reviews/issues/5595#issuecomment-1701418760. We need a tagged release created in the repository and then a related Zenodo archive DOI which is linked to this tagged release. The Zenodo archive should match the exact title of the paper, the author list and their information including ORCIDs.
