### 2023-08-31 #5595 @PieterjanRobbe
Great, thanks @jbytecode! I'll put the release on Zenodo and get a DOI asap.
