### 2023-08-31 #5595 @PieterjanRobbe
Hi @jbytecode, the version number is 2.2.4 and the DOI is 10.5281/zenodo.8306255. Thanks!
