### 2023-08-31 #5595 @jbytecode
@ziyiyin97, @shahmoradi - Many thanks for your efforts on reviewing this software and manuscript.

@PieterjanRobbe - Thank you for submitting your work to JOSS. Everyting looks good to me.

I am now recommending an acceptance. Our Editor-in-chief will make the final decision.
